Move interface implementations from test dependency resolver test file to helpers folder
